proj/api/serializers.py
from rest_framework import serializers
#from django.contrib.auth.models import User
from django.contrib.auth import get_user_model
from .models import UserStatistic, Game


class UserSerializer(serializers.ModelSerializer):

    class Meta:
        model = get_user_model()
        fields = ('username', 'password')
        extra_kwargs = {'password': {'write_only': True}}

    def create(self, validated_data):
        user = get_user_model().objects.create(
            username=validated_data['username'],
        )
        user.set_password(validated_data['password'])
        user.save()
        UserStatistic.objects.create(user=user)
        return user

class GameSerializer(serializers.ModelSerializer):
    class Meta:
        model = Game
        fields = ('id', 'size')